+++ Lecture6(GithubAPI)/Component/BodyComponent.js
@@ -15,20 +15,27 @@ const BodyComponent = () => {
     const githubUserNames = ["gavandivya", "akshaymarch7", "ap221882", "Bhallora", "aditifarkya", "it-abhishek2000", "deepak-kumar-dev", "soumyagangamwar", "shreerajjadeja", "nikitaj-57"]
 
     const getGithubApiData = async () => {
-        let githubdata = await Promise.all(githubUserNames.map(async (userName) => {
-            const UserInfo = await fetch(`https://api.github.com/users/${userName}`);
-            return await UserInfo.json();
-        }));
-        sessionStorage.setItem('teamData', JSON.stringify(githubdata));
-        return githubdata;
+        try {
+            let githubdata = await Promise.all(githubUserNames.map(async (userName) => {
+                const UserInfo = await fetch(`https://api.github.com/users/${userName}`);
+                return await UserInfo.json();
+            }));
+            sessionStorage.setItem('teamData', JSON.stringify(githubdata));
+            return githubdata;
+        }
+        catch (error) {
+            console.log(error);
+        }
     }
 
     useEffect(() => {
         if (!sessionStorage.getItem('teamData')) {
-            getGithubApiData().then((githubdata) => {
-                setLoading(false);
-                setDataFromAPI(githubdata);
-            });
+            getGithubApiData().then((githubdata) => setDataFromAPI(githubdata))
+                .catch((err) => {
+                    console.log(err);
+                    setNoRecord(true);
+                })
+                .finally(() => setLoading(false));
         }
         else {
             setDataFromAPI(JSON.parse(sessionStorage.getItem('teamData')));
