+++ Lecture6(GithubAPI)/utils/socialMediaIcon.js
@@ -1,0 +1,7 @@
+export const SocialMediaIcon = {
+    org: "https://img.icons8.com/plasticine/100/null/business-buildings.png",
+    loc: "https://img.icons8.com/fluency/48/null/marker.png",
+    logo: "https://images.weserv.nl/?url=https%3A%2F%2Fdrive.google.com%2Fuc%3Fexport%3Dview%26id%3D1ZN0OHfI7ngyLNIn0mDY36iFE2AewNMJ3&w=290&h=250&fit=cover&q=100",
+    title: "Brainy Fools"
+}
+
